+++ lib/Memmory/Memmory.h
@@ -24,8 +24,8 @@ public:
     static void save_daylight_saving(bool dlt_sv);
     static IPAddress get_server_ip();
     static void save_server_ip(const IPAddress &server_ip);
-    static unsigned long get_reminder_b_revision_no();
-    static void save_reminder_b_revision_no(unsigned long revision_no);
+    static uint32_t get_reminder_b_revision_no();
+    static void save_reminder_b_revision_no(uint32_t revision_no);
     static void save_wifi_cred(const String& ssid_, const String& pass_) ;
     static bool load_wifi_cred(String &WIFI_SSID, String &WIFI_PASS) ;
     static JsonDocument get_latest_Reminder(const String &time_string, JsonDocument &doc);
